refactor: extract ConfirmationDialog

+++ lib/admin/admin_screens/discount_screen.dart
@@ -2,6 +2,7 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
+import 'package:honey/widgets/custom_confirm_dialog.dart';
 import 'package:honey/widgets/title_appbar.dart';
 import 'package:honey/widgets/app_colors.dart';
 import 'package:honey/widgets/custom_button.dart';
@@ -100,56 +101,15 @@ class _DiscountScreenState extends State<DiscountScreen> {
           showDialog(
             context: context,
             builder: (BuildContext context) {
-              return Dialog(
-                backgroundColor: const Color.fromARGB(255, 27, 27, 27),
-                shape: RoundedRectangleBorder(
-                  borderRadius: BorderRadius.circular(15),
-                ),
-                child: Padding(
-                  padding: const EdgeInsets.all(30.0),
-                  child: Column(
-                    mainAxisSize: MainAxisSize.min,
-                    children: [
-                      const SizedBox(height: 15),
-                      const Text(
-                        'Видалити акцію на товар?',
-                        textAlign: TextAlign.center,
-                        style: TextStyle(
-                          fontSize: 16,
-                          fontFamily: 'MA',
-                        ),
-                      ),
-                      const SizedBox(height: 15),
-                      ButtonBar(
-                        alignment: MainAxisAlignment.spaceAround,
-                        children: [
-                          TextButton(
-                            onPressed: () async {
-                              Navigator.of(context).pop();
-                              Navigator.of(context).pop();
-                              _deleteDiscount();
-                            },
-                            child: const Text(
-                              'Так',
-                              style: TextStyle(
-                                  color: Colors.red, fontFamily: 'MA'),
-                            ),
-                          ),
-                          TextButton(
-                            onPressed: () {
-                              Navigator.of(context).pop();
-                            },
-                            child: const Text(
-                              'Скасувати',
-                              style: TextStyle(
-                                  color: Colors.white, fontFamily: 'MA'),
-                            ),
-                          ),
-                        ],
-                      ),
-                    ],
-                  ),
-                ),
+              return ConfirmationDialog(
+                title: 'Видалити акцію?',
+                confirmButtonText: 'Так',
+                cancelButtonText: 'Скасувати',
+                onConfirm: () {
+                  Navigator.of(context).pop();
+                  Navigator.of(context).pop();
+                  _deleteDiscount();
+                },
               );
             },
           );
