+++ lib/admin/admin_screens/product_edit_screen.dart
@@ -12,6 +12,7 @@ import 'package:honey/providers/product_model.dart';
 import 'package:honey/providers/products.dart';
 import 'package:honey/widgets/app_colors.dart';
 import 'package:honey/widgets/custom_button.dart';
+import 'package:honey/widgets/custom_confirm_dialog.dart';
 import 'discount_screen.dart';
 
 class ProductEditScreen extends StatefulWidget {
@@ -183,56 +184,16 @@ class _ProductEditScreenState extends State<ProductEditScreen> {
           showDialog(
             context: context,
             builder: (BuildContext context) {
-              return Dialog(
-                backgroundColor: const Color.fromARGB(255, 27, 27, 27),
-                shape: RoundedRectangleBorder(
-                  borderRadius: BorderRadius.circular(15),
-                ),
-                child: Padding(
-                  padding: const EdgeInsets.all(30.0),
-                  child: Column(
-                    mainAxisSize: MainAxisSize.min,
-                    children: [
-                      const SizedBox(height: 15),
-                      const Text(
-                        'Впевнені, що хочете видалити товар?',
-                        textAlign: TextAlign.center,
-                        style: TextStyle(
-                          fontSize: 16,
-                          fontFamily: 'MA',
-                        ),
-                      ),
-                      const SizedBox(height: 15),
-                      ButtonBar(
-                        alignment: MainAxisAlignment.spaceAround,
-                        children: [
-                          TextButton(
-                            onPressed: () async {
-                              Navigator.of(context).pop();
-                              deleteProductAndStorageImage(
-                                  context, widget.productId, _currentImageUrl!);
-                            },
-                            child: const Text(
-                              'Так',
-                              style: TextStyle(
-                                  color: Colors.red, fontFamily: 'MA'),
-                            ),
-                          ),
-                          TextButton(
-                            onPressed: () {
-                              Navigator.of(context).pop();
-                            },
-                            child: const Text(
-                              'Скасувати',
-                              style: TextStyle(
-                                  color: Colors.white, fontFamily: 'MA'),
-                            ),
-                          ),
-                        ],
-                      ),
-                    ],
-                  ),
-                ),
+              return ConfirmationDialog(
+                title: 'Впевнені, що хочете видалити товар?',
+                confirmButtonText: 'Так',
+                cancelButtonText: 'Скасувати',
+                onConfirm: () {
+                  Navigator.of(context).pop();
+                  Navigator.of(context).pop();
+                  deleteProductAndStorageImage(
+                      context, widget.productId, _currentImageUrl!);
+                },
               );
             },
           );
